requirements.txt
Flask
gunicorn
click
colorama
flask-mysql
Flask-Cors
pymysql
Flask-JWT-Extended
flask-bcrypt
waitress
